backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-socketio==5.11.1
aiohttp==3.9.3
python-multipart==0.0.6
python-dotenv==1.0.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
aiosqlite==0.20.0
requests==2.31.0